+++ src/components/DropdownSelect/index.stories.tsx
@@ -21,7 +21,17 @@ export const data: DropdownSelectContentItem[] = [
   { id: 'Item 6', checked: false, default: false },
   { id: 'Item 7', checked: false, default: false },
   { id: 'Item 8', checked: false, default: false },
-  { id: 'Item 9', checked: false, default: false },
+  {
+    id: 'Item 9',
+    checked: 'option-3',
+    default: false,
+    items: [
+      { id: 'option-1', label: 'Option 1' },
+      { id: 'option-2', label: 'Option 2' },
+      { id: 'option-3', label: 'Option 3' },
+      { id: 'option-4', label: 'Option 4' },
+    ],
+  },
   { id: 'Item 10', checked: false, default: false },
 ];
 
@@ -43,8 +53,6 @@ export const ListItem = styled.div`
   margin: 0;
   padding: 0.25rem 0;
   > b {
-    font-size: 0.875rem;
-    font-weight: 700;
     display: block;
     margin: 0;
     padding: 0;
@@ -64,11 +72,7 @@ const Template: Story<ISelectProps<DropdownSelectContentItem>> = (args) => {
       {...args}
       content={content}
       onChange={setContent}
-      itemRenderer={(item) => (
-        <ListItem>
-          <b>{item.id}</b>
-        </ListItem>
-      )}
+      itemRenderer={(item) => <ListItem>{item.id}</ListItem>}
       onConfirmSelection={(c: any) => alert(JSON.stringify(c))}
       onCancel={() => alert('cancel')}
       clearButtonText="Cancel"