+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.19.0] - 2021-07-14
+
+- Added DropdownSelect subitems
+- Fixed EditableLabel behavior
+
 ## [1.18.0] - 2021-07-12
 
 - Added Card component